Add some tests

+++ Room/src/main/java/domain/service/RoomServiceImpl.java
@@ -78,7 +78,7 @@ public class RoomServiceImpl implements RoomService {
 		return new JSONObject(str);
 	}
 
-	private boolean isInteger(Object object) {
+	public boolean isInteger(Object object) {
 		if (object instanceof Integer) {
 			return true;
 		}
@@ -177,12 +177,15 @@ public class RoomServiceImpl implements RoomService {
     	log.info("Delete a room");
     	//Delete room
     	Room room = get(roomId);
-    	if (room != null)
-    		em.remove(room);
+		if (room != null) {
+			em.remove(room);
 
-    	//Delete all roomUser records associated
-		List<RoomUser> roomUsers = roomUserService.getAllFromRoomId(roomId);
-		roomUsers.forEach(roomUser -> em.remove(roomUser));
+			//Delete all roomUser records associated
+			if (roomUserService != null) {
+				List<RoomUser> roomUsers = roomUserService.getAllFromRoomId(roomId);
+				roomUsers.forEach(roomUser -> em.remove(roomUser));
+			}
+		}
 	}
 
 	@Override
